Invisibles23/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudinary==1.33.0
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==4.1.0
django-environ==0.10.0
django-filter==23.2
django-js-asset==2.0.0
djlint==1.19.16
EditorConfig==0.12.3
gunicorn==20.1.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==3.4
jsbeautifier==1.14.7
mailchimp-marketing @ git+https://github.com/mailchimp/mailchimp-marketing-python.git@3305fa45b3f436767a539c5fba9cb2b0a083d761
pathspec==0.11.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.31.0
six==1.16.0
sqlparse==0.4.3
stripe==5.4.0
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.16
whitenoise==6.5.0
